windows/requirements.txt
requests>=2.31.0
flask>=3.0.0
pillow>=10.0.0
pystray>=0.19.5
APScheduler>=3.10.0
pyinstaller>=6.0.0
